Rest API

+++ epictaskapi/src/main/java/br/com/fiap/epictaskapi/controller/TaskController.java
@@ -1,9 +1,9 @@
 package br.com.fiap.epictaskapi.controller;
 
 import java.util.List;
-
 import javax.validation.Valid;
 
+import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -41,9 +42,35 @@ public class TaskController {
 
     @DeleteMapping("{id}")
     public ResponseEntity<Object> destroy(@PathVariable Long id){
+        var optional = service.getById(id);
+
+        if(optional.isEmpty())
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+
         service.remove(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
+
+    @GetMapping("{id}")
+    public ResponseEntity<Task> show(@PathVariable Long id){
+        return ResponseEntity.of(service.getById(id));
+    }
+
+    @PutMapping("{id}")
+    public ResponseEntity<Task> update(@PathVariable Long id, @RequestBody @Valid Task newTask){
+        var optional = service.getById(id);
+
+        if(optional.isEmpty())
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+
+        var task = optional.get();
+        BeanUtils.copyProperties(newTask, task);
+        task.setId(id);
+
+        service.save(task);
+        return ResponseEntity.ok(task);
+    }
+
